+++ Magic/CHANGELOG.md
@@ -7,6 +7,8 @@
    - These are all old, before the action system, and have been deprecated for years
    - They will show in logs in startup
    - They WILL be removed in future versions!
+ - Add ModifyVariable action, adding concept of user-defined variables that can be used in spell parameters
+   - TODO: actions need to be re-prepared after variables change
  - Fix potter wand boxes not properly turning into a wand (icon wasn't working)
  - Fix missing IMPORTANT headers in defaults files
  - Add TakeCosts action